FloodKal/requirements.txt
appdirs==1.4.3
APScheduler==3.4.0
asn1crypto==0.22.0
beautifulsoup4==4.3.2
cffi==1.10.0
cryptography==1.8.1
cssselect==0.9.1
DistributedLock==1.2
dj-database-url==0.4.1
Django==1.11.3
django-annoying==0.10.3
django-appconf==1.0.2
django-bootstrap3==9.0.0
django-compressor==2.2
django-configurations==2.0
django-crispy-forms==1.6.1
django-debug-toolbar==1.8
django-extensions==1.7.5
django-geojson==2.10.0
django-jsonview==0.4.3
django-leaflet==0.22.0
django-lineage==0.2.0
django-m2m-history==0.3.6
django-oauth-tokens==0.6.3
django-picklefield==1.0.0
django-tag-parser==2.1
django-taggit==0.22.1
django-twitter-api==0.2.10
-e git+https://github.com/michaelbrooks/django-twitter-stream.git@0b44f5e273b84ff93a2023ef504b334fe201232c#egg=django_twitter_stream
djangorestframework==3.5.3
dominate==2.1.17
enum34==1.1.6
feedfinder2==0.0.1
feedparser==5.1.3
funcsigs==1.0.2
futures==3.1.1
geograpy==0.3.7
gunicorn==19.6.0
idna==2.5
ipaddress==1.0.18
jellyfish==0.5.6
jieba==0.35
jsonfield==2.0.2
lxml==3.3.5
mongoengine==0.15.0
mysqlclient==1.3.4
ndg-httpsclient==0.4.2
newspaper==0.1.0.7
nltk==3.2.5
numpy==1.13.3
oauthlib==2.0.2
olefile==0.44
packaging==16.8
Pillow==2.5.1
psycopg2==2.6.2
pyasn1==0.2.3
pycountry==17.9.23
pycparser==2.17
pymongo==3.4.0
PyMySQL==0.7.11
pyOpenSSL==16.2.0
pyparsing==2.2.0
pysocks==1.6.7
python-dateutil==2.4.0
python-memcached==1.58
pytz==2017.3
PyYAML==3.11
rcssmin==1.0.6
requests==2.6.0
requests-oauthlib==0.8.0
requirements==0.1
rjsmin==1.0.12
simplejson==3.11.1
six==1.7.3
South==1.0.2
sqlparse==0.2.4
swapper==0.1.1
tldextract==1.5.1
tweepy==3.3.0
twitter-monitor==0.3.0
tzlocal==1.4
virtualenv==15.1.0
whitenoise==2.0.6
